Update interface to have dockestServices declare their dependents rather than the inverse
