+++ packages/dockest/src/run/bootstrap/index.ts
@@ -1,11 +1,8 @@
 import { configureLogger } from './configureLogger'
-import { createRunners } from './createRunners'
-import { injectDependees } from './injectDependees'
 import { mergeComposeFiles } from './mergeComposeFiles'
 import { parseComposeFile } from './parseComposeFile'
 import { setupExitHandler } from './setupExitHandler'
-import { validateRunners } from './validateRunners'
-import { validateRunnersWithDependsOn } from './validateRunnersWithDependsOn'
+import { transformDockestServicesToRunners } from './transformDockestServicesToRunners'
 import { writeComposeFile } from './writeComposeFile'
 import { DockestConfig } from '../../@types'
 
@@ -16,11 +13,7 @@ export const bootstrap = async (config: DockestConfig) => {
   const { composeFileAsObject } = parseComposeFile(mergedComposeFiles)
   writeComposeFile(mergedComposeFiles, composeFileAsObject)
 
-  const { runners, runnersWithDependsOn } = createRunners(config, composeFileAsObject)
-  validateRunners(runners)
-  validateRunnersWithDependsOn(runnersWithDependsOn)
-  injectDependees(runners, runnersWithDependsOn)
-  config.$.runners = runners
+  transformDockestServicesToRunners(config, composeFileAsObject)
 
   configureLogger(config)
 }
