+++ packages/dockest/src/run/debugMode.ts
@@ -11,9 +11,7 @@ export const debugMode = async (config: DockestConfig) => {
   if (debug || process.argv.includes('dev') || process.argv.includes('debug')) {
     Logger.info(`Debug mode enabled, containers are kept running and Jest will not run.`)
 
-    Object.values(runners).forEach(runner =>
-      Logger.info(`[${runner.dockestService.serviceName}]: ${runner.containerId}`),
-    )
+    Object.values(runners).forEach(runner => Logger.info(`[${runner.serviceName}]: ${runner.containerId}`))
 
     await sleep(1000 * 60 * 60 * 24)
   }
