+++ packages/dockest/src/run/waitForServices/healthcheck.ts
@@ -6,11 +6,7 @@ import { Runner } from '../../@types'
 const logPrefix = '[Check Responsiveness]'
 
 export const checkResponsiveness = async (runner: Runner) => {
-  const {
-    containerId,
-    dockestService: { healthchecks = [] },
-    logger,
-  } = runner
+  const { containerId, healthchecks = [], logger } = runner
   const responsivenessTimeout = 30
 
   await Promise.all(
