+++ packages/dockest/src/run/waitForServices/runRunnerCommands.ts
@@ -4,9 +4,7 @@ import { Runner } from '../../@types'
 const logPrefix = '[Dockest Service Commands]'
 
 export const runRunnerCommands = async (runner: Runner) => {
-  const {
-    dockestService: { commands = [] },
-  } = runner
+  const { commands = [] } = runner
 
   for (const command of commands) {
     await execaWrapper(command, { runner, logPrefix, logStdout: true })
